Fix build on aarch64
`ash::vk::ApplicationInfo` [1] actually uses `*const c_char` as a type
for its string fields, and the underlying type of `libc::c_char` varies on
different architectures.

Because of this, the application currently builds fine on x64 (where
`c_char` is `i8`) but fails on aarch64 (where `c_char` is `u8` instead).

[1]: https://docs.rs/ash/latest/ash/vk/struct.ApplicationInfo.html